Removed SaveableContainer, SaveableObject is used instead (data that should not be saved should be placed to non-saveable sub-conainer, instead of placing data that should be saved into saveable sub-container. This removes the need to have XXXData class for every XXX class.) Files are not removet yet in this commit.
Added test generation of world, realm zone and room. It's not working as intended yet.

Added base of support for GameEntities to be able to contain other game entities. It's not working yet, thanks to bug in V8 engine disallowing using super. in async functions. Some workaround will be necessary.